requirements.txt
fastapi==0.65.2
uvicorn==0.17.4
mysql-connector-python==8.0.27
